requirements.txt
telethon==1.34.0
asyncio
python-dotenv==1.0.0 